GeekTG 3.0

+++ friendly-telegram/utils.py
@@ -39,16 +39,22 @@ def get_args(message):
         message = message.message
     except AttributeError:
         pass
+
     if not message:
         return False
+
     message = message.split(maxsplit=1)
+
     if len(message) <= 1:
         return []
+
     message = message[1]
+
     try:
         split = shlex.split(message)
     except ValueError:
         return message  # Cannot split, let's assume that it's just one long message
+
     return list(filter(lambda x: len(x) > 0, split))
 
 
@@ -58,11 +64,15 @@ def get_args_raw(message):
         message = message.message
     except AttributeError:
         pass
+
     if not message:
         return False
+
     args = message.split(maxsplit=1)
+
     if len(args) > 1:
         return args[1]
+
     return ""
 
 
@@ -70,6 +80,7 @@ def get_args_split_by(message, sep):
     """Split args with a specific sep"""
     raw = get_args_raw(message)
     mess = raw.split(sep)
+
     return [section.strip() for section in mess if section]
 
 
@@ -108,19 +119,24 @@ async def get_user(message):
         return await message.client.get_entity(message.sender_id)
     except ValueError:  # Not in database. Lets go looking for them.
         logging.debug("user not in session cache. searching...")
+
     if isinstance(message.to_id, PeerUser):
         try:
             await message.client.get_dialogs()
-        except telethon.rpcerrorlist.BotMethodInvalid:  # TODO: unable find BotMethodInvalid
+        except telethon.rpcerrorlist.BotMethodInvalid:
             return None
+
         return await message.client.get_entity(message.sender_id)
+
     if isinstance(message.to_id, (PeerChannel, PeerChat)):
         async for user in message.client.iter_participants(message.to_id, aggressive=True):
             if user.id == message.sender_id:
                 return user
-        logging.error("WTF! user isn't in the group where they sent the message")
+
+        logging.critical("WTF! user isn't in the group where they sent the message")
         return None
-    logging.error("WTF! to_id is not a user, chat or channel")
+
+    logging.critical("WTF! to_id is not a user, chat or channel")
     return None
 
 
@@ -141,17 +157,20 @@ def censor(obj, to_censor=None, replace_with="redacted_{count}_chars"):  # pylin
     """May modify the original object, but don't rely on it"""
     if to_censor is None:
         to_censor = ["phone"]
+
     for k, v in vars(obj).items():
         if k in to_censor:
             setattr(obj, k, replace_with.format(count=len(v)))
         elif k[0] != "_" and hasattr(v, "__dict__"):
             setattr(obj, k, censor(v, to_censor, replace_with))
+
     return obj
 
 
 def relocate_entities(entities, offset, text=None):
     """Move all entities by offset (truncating at text)"""
     length = len(text) if text is not None else 0  # TODO: refactor about text=None
+
     for ent in entities.copy() if entities else ():
         ent.offset += offset
         if ent.offset < 0:
@@ -161,6 +180,7 @@ def relocate_entities(entities, offset, text=None):
             ent.length = length - ent.offset
         if ent.length <= 0:
             entities.remove(ent)
+
     return entities
 
 
@@ -169,10 +189,12 @@ def _fix_entities(ent, cont_msg, initial=False):
         if not initial:
             entity.offset -= len(cont_msg)
             entity.length += len(cont_msg)
+
         if entity.offset + entity.length < 4096:
             entity.offset = 0
             entity.length = 0  # It's useless
             continue
+
         already_done = max(4096 - entity.offset, 0)
         entity.offset = max(entity.offset - 4096, 0)
         entity.length -= already_done
@@ -186,23 +208,28 @@ async def answer(message, response, **kwargs):
         message = message[0]
     else:
         delete_job = None
+
     if await message.client.is_bot() and isinstance(response, str) and len(response) > 4096:
         kwargs.setdefault("asfile", True)
+
     kwargs.setdefault("link_preview", False)
+
     edit = message.out
+
     if not edit:
         kwargs.setdefault("reply_to", message.reply_to_msg_id if await message.get_reply_message() else message.id)
+
     parse_mode = telethon.utils.sanitize_parse_mode(kwargs.pop("parse_mode", message.client.parse_mode))
+
     if isinstance(response, str) and not kwargs.pop("asfile", False):
         txt, ent = parse_mode.parse(response)
-        logging.debug(txt)
-        logging.debug(ent)
 
         if len(txt) >= 8192:
             file = io.BytesIO(txt.encode('utf-8'))
             file.name = "command_result.txt"
             await message.client.send_file(message.to_id, file, caption="<b>📤 Command output seems to be too long, so it's sent in file.</b>")
-            await message.delete()
+            if message.out:
+                await message.delete()
             return
 
         ret = [await (message.edit if edit else message.respond)(txt[:4096], parse_mode=lambda t: (t, ent), **kwargs)]
@@ -231,11 +258,13 @@ async def answer(message, response, **kwargs):
     else:
         if isinstance(response, bytes):
             response = io.BytesIO(response)
+
         if isinstance(response, str):
             response = io.BytesIO(response.encode("utf-8"))
-        name = kwargs.pop("filename", None)
-        if name:
+
+        if name := kwargs.pop("filename", None):
             response.name = name
+
         if message.media is not None and edit:
             await message.edit(file=response, **kwargs)
         else:
@@ -245,9 +274,11 @@ async def answer(message, response, **kwargs):
             ret = (await message.client.send_file(message.chat_id, response,
                                                   **kwargs),)
             await new.delete()
+
     if delete_job:
         await delete_job
-    return ret  # TODO: refactor
+
+    return ret
 
 
 async def get_target(message, arg_no=0):
@@ -263,6 +294,7 @@ async def get_target(message, arg_no=0):
         user = message.to_id.user_id
     else:
         return None
+
     try:
         ent = await message.client.get_entity(user)
     except ValueError:
